dashboard/requirements.txt
streamlit
streamz
pandas
matplotlib
kafka-python
sqlalchemy
psycopg2-binary
plotly